fixup: put flairVersion on the bootstrap RESPONSE, not the request body
- Extract resolveVersion() into shared resources/version.ts instead of
  leaving a third copy (Presence.ts / AdminInstance.ts still carry their
  own copies — migrating those is a follow-up).
- Attach flairVersion to the returned response object, not the delegated
  request body. BootstrapMemories.post builds its response as an explicit
  object literal and never echoes the request, so the field was silently
  dropped. The documented invariant (plain bootstrap delegates a
  byte-identical body) is now preserved.
- Test asserts on body.result.structuredContent.flairVersion (the
  response the agent actually sees), and also asserts the delegated
  request body does NOT carry flairVersion.
